fix(catalog): remove trailing slashes of all URLS.
sphinx appends `/objects.inv` to the URLs given to it. In sphinx>=8.0
this ends up doubling the `/` if the URL ends with one already.
Our tests failed as we read the output of sphinx and were expecting
"clean" URLs.

This ensures every URL given to sphinx by `Catalog` does not end with a slash.