fix(catalog): remove trailing slashes of all URLS given to sphinx. (#6)
* fix(catalog): remove trailing slashes of all URLS.

sphinx appends `/objects.inv` to the URLs given to it. In sphinx>=8.0
this ends up doubling the `/` if the URL ends with one already.
Our tests failed as we read the output of sphinx and were expecting
"clean" URLs.

This ensures every URL given to sphinx by `Catalog` does not end with a slash.

* fix: detect conflict with/without URL trailing '/'

When addinga new URL to an intersphinx mapping, detect a conflict even
when one of the URL ends with a `/` and not the other.

Prevents duplicate URLs where only the trailing `/` differs.

* chore(deps): update the pixi lock file.

Makes the tests on GitHub Action run against sphinx 8.0.2.